Support for es6 imports and straight browser usage
- Convert build to webpack
- Use webpack-dev-server for development
- Replace karma with jest for tests
- Replace gulp with npm for scripts/builds
