+++ .babelrc
@@ -1,0 +1,8 @@
+{
+    "presets": [["env", {"modules": false}]],
+    "env": {
+        "test": {
+            "presets": ["env"]
+        }
+    }
+}
