--- bower.json
@@ -1,32 +1,0 @@
-{
-  "name": "json-schema-view-js",
-  "version": "1.0.0",
-  "authors": [
-    "Mohsen Azimi <me@azimi.me>"
-  ],
-  "main": [
-    "dist/bundle.js",
-    "dist/style.css"
-  ],
-  "moduleType": [
-    "amd",
-    "es6",
-    "globals",
-    "node"
-  ],
-  "keywords": [
-    "json"
-  ],
-  "license": "MIT",
-  "homepage": "mohsen1/json-schema-view-js",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "dependencies": {
-    "json-formatter-js": "~0.3.0"
-  }
-}
