+++ package.json
@@ -4,8 +4,13 @@
   "description": "JSON Schema View core library ",
   "main": "dist/bundle.js",
   "scripts": {
-    "test": "gulp test --singleRun=true",
-    "build": "gulp"
+    "start": "webpack-dev-server --open",
+    "test": "jest",
+    "build": "rm -rf ./dist && npm run build:prod && npm run build:dev",
+    "build:dev": "webpack --env.dev",
+    "build:prod": "webpack --env.production",
+    "prepublish": "npm run build",
+    "precommit": "lint-staged && npm test"
   },
   "repository": {
     "type": "git",
@@ -21,29 +26,56 @@
   },
   "homepage": "https://github.com/mohsen1/json-schema-view-js#readme",
   "dependencies": {
-    "json-formatter-js": "^2.2.0"
+    "json-formatter-js": "^2.2.0",
+    "vinyl": "^2.1.0"
   },
   "devDependencies": {
-    "babel": "^5.8.21",
-    "babelify": "^6.2.0",
-    "browserify": "^11.0.1",
-    "chai": "^3.2.0",
-    "gulp": "^3.9.0",
-    "gulp-connect": "^2.2.0",
-    "gulp-header": "^1.7.1",
-    "gulp-jshint": "^1.11.2",
-    "gulp-less": "^3.0.3",
-    "gulp-minify-css": "^1.2.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-uglify": "^1.4.0",
-    "karma": "^0.13.9",
-    "karma-babel-preprocessor": "^5.2.1",
-    "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^0.2.0",
-    "karma-firefox-launcher": "^0.1.6",
-    "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.1.1",
-    "minimist": "^1.2.0",
-    "mocha": "^2.2.5"
+    "babel-core": "^6.26.0",
+    "babel-jest": "^21.2.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.1",
+    "css-loader": "^0.28.7",
+    "eslint": "^4.9.0",
+    "eslint-loader": "^1.9.0",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "html-webpack-plugin": "^2.30.1",
+    "husky": "^0.14.3",
+    "jest": "^21.2.1",
+    "jest-css-modules": "^1.1.0",
+    "less": "^3.0.0-alpha.3",
+    "less-loader": "^4.0.5",
+    "lint-staged": "^4.3.0",
+    "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "style-loader": "^0.19.0",
+    "uglifyjs-webpack-plugin": "^1.0.0-beta.3",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.3",
+    "webpack-stream": "^4.0.0"
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx}"
+    ],
+    "setupFiles": [],
+    "testMatch": [
+      "<rootDir>/src/**/__tests__/**/*.js?(x)",
+      "<rootDir>/src/**/?(*.)(spec|test).js?(x)",
+      "<rootDir>/test/**/?(*.)js?(x)"
+    ],
+    "testEnvironment": "jsdom",
+    "testURL": "http://localhost",
+    "transform": {
+      "^.+\\.(js|jsx)$": "<rootDir>/node_modules/babel-jest",
+      ".+\\.(css|styl|less|sass|scss)$": "<rootDir>/node_modules/jest-css-modules"
+    },
+    "transformIgnorePatterns": [
+      "[/\\\\]node_modules[/\\\\](?!lodash-es).+\\.(js|jsx)$"
+    ]
+  },
+  "lint-staged": {
+    "*.js": [
+      "eslint --fix",
+      "git add"
+    ]
   }
 }
